+++ src/api/index.ts
@@ -1,10 +1,10 @@
-import { Network, TokenList, WalletApi, Erc20Api } from 'types'
+import { Network, TokenList } from 'types'
 import { WalletApiMock } from './wallet/WalletApiMock'
-import WalletApiImpl from './wallet/WalletApiImpl'
+import { WalletApiImpl, WalletApi } from './wallet/WalletApiImpl'
 import { TokenListApiImpl } from './tokenList/TokenListApiImpl'
 import { TokenListApiMock } from './tokenList/TokenListApiMock'
 import { Erc20ApiMock } from './erc20/Erc20ApiMock'
-import { Erc20ApiImpl } from './erc20/Erc20ApiImpl'
+import { Erc20ApiImpl, Erc20Api } from './erc20/Erc20ApiImpl'
 import { DepositApiMock } from './exchange/DepositApiMock'
 import { DepositApiImpl, DepositApi } from './exchange/DepositApiImpl'
 import { ExchangeApiImpl, ExchangeApi } from './exchange/ExchangeApiImpl'
