+++ src/types/index.ts
@@ -1,6 +1,5 @@
 import BN from 'bn.js'
-import Web3Connect from 'web3connect'
-import { Provider } from '@gnosis.pm/dapp-ui'
+import { TransactionReceipt } from 'web3-core'
 
 export type Command = () => void
 export type Mutation<T> = (original: T) => T
@@ -57,50 +56,6 @@ export interface TxOptionalParams {
 
 export type Receipt = TransactionReceipt
 
-export interface EpochTokenLocker extends Contract {
-  clone(): EpochTokenLocker
-
-  methods: {
-    hasCreditedBalance(user: string, token: string, batchId: number | string | BN): TransactionObject<boolean>
-
-    BATCH_TIME(): TransactionObject<string>
-
-    deposit(token: string, amount: number | string | BN): TransactionObject<void>
-
-    requestWithdraw(token: string, amount: number | string | BN): TransactionObject<void>
-
-    withdraw(owner: string, token: string): TransactionObject<void>
-
-    getPendingDepositAmount(user: string, token: string): TransactionObject<string>
-
-    getPendingDepositBatchNumber(user: string, token: string): TransactionObject<string>
-
-    getPendingWithdrawAmount(user: string, token: string): TransactionObject<string>
-
-    getPendingWithdrawBatchNumber(user: string, token: string): TransactionObject<string>
-
-    // eslint-disable-next-line @typescript-eslint/adjacent-overload-signatures
-    withdraw(
-      {
-        userAddress,
-        tokenAddress,
-      }: {
-        userAddress: string
-        tokenAddress: string
-      },
-      txOptionalParams?: TxOptionalParams,
-    ): Promise<Receipt>
-
-    getCurrentBatchId(): TransactionObject<string>
-
-    getSecondsRemainingInBatch(): TransactionObject<string>
-
-    getBalance(user: string, token: string): TransactionObject<string>
-
-    hasValidWithdrawRequest(user: string, token: string): TransactionObject<boolean>
-  }
-}
-
 export interface Order {
   buyTokenId: number
   sellTokenId: number
@@ -124,233 +79,4 @@ export interface PlaceOrderParams {
   validUntil: number
   buyAmount: BN
   sellAmount: BN
-}
-
-export interface WalletInfo {
-  isConnected: boolean
-  userAddress?: string
-  networkId?: number
-}
-
-export type ProviderInfo = ReturnType<typeof Web3Connect.getProviderInfo>
-
-export interface WalletApi {
-  isConnected(): boolean
-  connect(givenProvider?: Provider): Promise<boolean>
-  disconnect(): Promise<void>
-  getAddress(): Promise<string>
-  getBalance(): Promise<BN>
-  getNetworkId(): Promise<number>
-  addOnChangeWalletInfo(callback: (walletInfo: WalletInfo) => void, trigger?: boolean): Command
-  removeOnChangeWalletInfo(callback: (walletInfo: WalletInfo) => void): void
-  getProviderInfo(): ProviderInfo
-}
-
-import { Contract } from 'web3-eth-contract'
-import { PromiEvent, TransactionConfig, TransactionReceipt } from 'web3-core'
-
-type CallTxOptions = Pick<TransactionConfig, 'from' | 'gas' | 'gasPrice'>
-type SendTxOptions = Pick<TransactionConfig, 'from' | 'gas' | 'gasPrice' | 'value'>
-type EstimateGasTxOptions = Pick<TransactionConfig, 'from' | 'gas' | 'value'>
-
-type CallTxCallback = <T>(error: Error | null, result: T) => void
-type SendTxCallback = (error: Error | null, transactionHash: string) => void
-type EstimateGasTxCallback = (error: Error | null, gasAmount: number) => void
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-interface TransactionObject<T, U extends any[] = []> {
-  arguments: U
-
-  call(callback: CallTxCallback): Promise<T>
-  call(tx?: CallTxOptions, callback?: CallTxCallback): Promise<T>
-
-  send(callback: SendTxCallback): PromiEvent<TransactionReceipt>
-  send(tx?: SendTxOptions, callback?: SendTxCallback): PromiEvent<TransactionReceipt>
-
-  estimateGas(callback: EstimateGasTxCallback): Promise<number>
-  estimateGas(tx?: EstimateGasTxOptions, callback?: EstimateGasTxCallback): Promise<number>
-
-  encodeABI(): string
-}
-
-export interface ERC20 extends Contract {
-  clone(): ERC20
-
-  methods: {
-    totalSupply(): TransactionObject<string>
-    decimals(): TransactionObject<string>
-    symbol(): TransactionObject<string>
-    name(): TransactionObject<string>
-
-    balanceOf(owner: string): TransactionObject<string, [string]>
-
-    allowance(owner: string, spender: string): TransactionObject<string, [string, string]>
-
-    approve(spender: string, value: number | string | BN): TransactionObject<boolean, [string, string | number | BN]>
-
-    transfer(to: string, value: number | string | BN): TransactionObject<boolean, [string, string | number | BN]>
-
-    transferFrom(
-      from: string,
-      to: string,
-      value: number | string | BN,
-    ): TransactionObject<boolean, [string, string, string | number | BN]>
-  }
-}
-
-export interface BatchExchange extends Contract {
-  clone(): BatchExchange
-
-  methods: EpochTokenLocker['methods'] & {
-    getSecondsRemainingInBatch(): TransactionObject<string>
-
-    feeDenominator(): TransactionObject<string>
-
-    getPendingWithdrawAmount(user: string, token: string): TransactionObject<string, [string, string]>
-
-    requestWithdraw(
-      token: string,
-      amount: number | string | BN,
-    ): TransactionObject<void, [string, number | string | BN]>
-
-    getPendingDepositAmount(user: string, token: string): TransactionObject<string, [string, string]>
-
-    deposit(token: string, amount: number | string | BN): TransactionObject<void, [string, number | string | BN]>
-
-    getPendingWithdrawBatchNumber(user: string, token: string): TransactionObject<string, [string, string]>
-
-    TOKEN_ADDITION_FEE_IN_OWL(): TransactionObject<string>
-
-    feeToken(): TransactionObject<string>
-
-    currentPrices(arg0: number | string | BN): TransactionObject<string, [number | string | BN]>
-
-    orders(
-      arg0: string,
-      arg1: number | string | BN,
-    ): TransactionObject<
-      {
-        buyToken: BN
-        sellToken: BN
-        validFrom: BN
-        validUntil: BN
-        isSellOrder: boolean
-        priceNumerator: BN
-        priceDenominator: BN
-        usedAmount: BN
-        0: BN
-        1: BN
-        2: BN
-        3: BN
-        4: BN
-        5: BN
-        6: BN
-      },
-      [string, number | string | BN]
-    >
-
-    numTokens(): TransactionObject<string>
-
-    lastCreditBatchId(arg0: string, arg1: string): TransactionObject<string>
-
-    latestSolution(): TransactionObject<{
-      batchId: string
-      solutionSubmitter: string
-      feeReward: string
-      objectiveValue: string
-      0: string
-      1: string
-      2: string
-      3: string
-    }>
-
-    getBalance(user: string, token: string): TransactionObject<string, [string, string]>
-
-    getCurrentBatchId(): TransactionObject<string>
-
-    requestFutureWithdraw(
-      token: string,
-      amount: number | string | BN,
-      batchId: number | string | BN,
-    ): TransactionObject<void, [string, number | string | BN, number | string | BN]>
-
-    hasValidWithdrawRequest(user: string, token: string): TransactionObject<boolean, [string, string]>
-
-    MAX_TOKENS(): TransactionObject<string>
-
-    getPendingDepositBatchNumber(user: string, token: string): TransactionObject<string, [string, string]>
-
-    withdraw(user: string, token: string): TransactionObject<void, [string, string]>
-
-    MAX_TOUCHED_ORDERS(): TransactionObject<string>
-
-    addToken(token: string): TransactionObject<void, [string]>
-
-    placeValidFromOrders(
-      buyTokens: (number | string | BN)[],
-      sellTokens: (number | string | BN)[],
-      validFroms: (number | string | BN)[],
-      validUntils: (number | string | BN)[],
-      buyAmounts: (number | string | BN)[],
-      sellAmounts: (number | string | BN)[],
-    ): TransactionObject<
-      string[],
-      [
-        (number | string | BN)[],
-        (number | string | BN)[],
-        (number | string | BN)[],
-        (number | string | BN)[],
-        (number | string | BN)[],
-        (number | string | BN)[],
-      ]
-    >
-
-    placeOrder(
-      buyToken: number | string | BN,
-      sellToken: number | string | BN,
-      validUntil: number | string | BN,
-      buyAmount: number | string | BN,
-      sellAmount: number | string | BN,
-    ): TransactionObject<
-      string,
-      [number | string | BN, number | string | BN, number | string | BN, number | string | BN, number | string | BN]
-    >
-
-    cancelOrder(ids: (number | string | BN)[]): TransactionObject<void, [(number | string | BN)[]]>
-
-    freeStorageOfOrder(ids: (number | string | BN)[]): TransactionObject<void, [(number | string | BN)[]]>
-
-    submitSolution(
-      batchIndex: number | string | BN,
-      claimedObjectiveValue: number | string | BN,
-      owners: string[],
-      orderIds: (number | string | BN)[],
-      volumes: (number | string | BN)[],
-      prices: (number | string | BN)[],
-      tokenIdsForPrice: (number | string | BN)[],
-    ): TransactionObject<
-      string,
-      [
-        number | string | BN,
-        number | string | BN,
-        string[],
-        (number | string | BN)[],
-        (number | string | BN)[],
-        (number | string | BN)[],
-        (number | string | BN)[],
-      ]
-    >
-
-    tokenAddressToIdMap(addr: string): TransactionObject<string, [string]>
-
-    tokenIdToAddressMap(id: number | string | BN): TransactionObject<string, [number | string | BN]>
-
-    hasToken(addr: string): TransactionObject<boolean, [string]>
-
-    getEncodedAuctionElements(): TransactionObject<string>
-
-    acceptingSolutions(batchIndex: number | string | BN): TransactionObject<boolean, [number | string | BN]>
-
-    getCurrentObjectiveValue(): TransactionObject<string>
-  }
 }
